src/app/books/book-list/book-list.component.spec.ts
import { async, ComponentFixture, TestBed } from '@angular/core/testing';

import { BookListComponent } from './book-list.component';

describe('BookListComponent', () => {
  let component: BookListComponent;
  let fixture: ComponentFixture<BookListComponent>;

  beforeEach(async(() => {
    TestBed.configureTestingModule({
      declarations: [ BookListComponent ]
    })
    .compileComponents();
  }));

  beforeEach(() => {
    fixture = TestBed.createComponent(BookListComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    expect(component).toBeTruthy();
  });
  
  it('should have 3 books', () => {
    expect(false).toBeTruthy();
  });
  it('should display the title of each book', () => {
    expect(false).toBeTruthy();
  });
  it('should link to the details page of each book', () => {
    expect(false).toBeTruthy();
  });
});
